+++ Models/DTO/InventoryManagement/BulkUploadItemsResponse.cs
@@ -9,13 +9,13 @@
         public string BarCode { get; set; }
         public string Measurement { get; set; }
         public string UnitName { get; set; }
-        public double PurchaseRate { get; set; }
-        public double SalesRate { get; set; }
-        public double DiscountAmount { get; set; }
+        public decimal PurchaseRate { get; set; }
+        public decimal SalesRate { get; set; }
+        public decimal DiscountAmount { get; set; }
         public bool IsDiscountInPercentage { get; set; }
         public string CategoryName { get; set; }
         public string SubCategoryName { get; set; }
-        public double MinimumQuantity { get; set; }
+        public decimal MinimumQuantity { get; set; }
         public string BrandName { get; set; }
         public string ColorName { get; set; }
         public string SizeName { get; set; }
