+++ Models/DTO/InventoryManagement/InvGrnDetailsDTO.cs
@@ -7,10 +7,10 @@
         //public string PoNo { get; set; }
         public int ItemId { get; set; }
         public string BatchNo { get; set; }
-        //public double? OrderedQuantity { get; set; }
-        public double ReceivedQuantity { get; set; }
-        public double Rate { get; set; }
-        public double SubTotal => (ReceivedQuantity) * Rate;
+        //public decimal? OrderedQuantity { get; set; }
+        public decimal ReceivedQuantity { get; set; }
+        public decimal Rate { get; set; }
+        public decimal SubTotal => (ReceivedQuantity) * Rate;
 
         public virtual InvGrnMasterDto Grn { get; set; }
         public virtual InvItemDto Item { get; set; }
