+++ Models/DTO/InventoryManagement/InvGrrnDetailsDTO.cs
@@ -6,10 +6,10 @@ namespace Models.DTO.InventoryManagement
         public int GrrnId { get; set; }
         public int ItemId { get; set; }
         public string BatchNo { get; set; }
-        public double ReturnQuantity { get; set; }
-        public double Rate { get; set; }
+        public decimal ReturnQuantity { get; set; }
+        public decimal Rate { get; set; }
         public virtual InvGrrnMasterDto Grrn { get; set; }
         public virtual InvItemDto Item { get; set; }
-        public double SubTotal => (ReturnQuantity) * Rate;
+        public decimal SubTotal => (ReturnQuantity) * Rate;
     }
 }
