+++ Models/DTO/InventoryManagement/InvPhysicalInventoryItemDTO.cs
@@ -11,11 +11,11 @@ namespace Models.DTO.InventoryManagement
         public int? BarCodeId { get; set; }
         public int? VendorId { get; set; }
         public DateTime? ExpiryDate { get; set; }
-        public double? Quantity { get; set; }
-        public double? RemainingQuantity { get; set; }
+        public decimal? Quantity { get; set; }
+        public decimal? RemainingQuantity { get; set; }
         public int? TaxId { get; set; }
-        public double? PurchaseRate { get; set; }
-        public double? SalesRate { get; set; }
+        public decimal? PurchaseRate { get; set; }
+        public decimal? SalesRate { get; set; }
 
         public virtual InvItemDto Item { get; set; }
         public virtual InvItemBarCodeDto BarCode { get; set; }
