+++ Models/DTO/InventoryManagement/InvPoDetailsDTO.cs
@@ -5,14 +5,14 @@ namespace Models.DTO.InventoryManagement
     {
         public int PoId { get; set; }
         public int ItemId { get; set; }
-        public double RequestedQuantity { get; set; }
-        public double Rate { get; set; }
+        public decimal RequestedQuantity { get; set; }
+        public decimal Rate { get; set; }
 
         public virtual InvItemDto Item { get; set; }
         public virtual InvPoMasterDto Po { get; set; }
 
 
         //dto
-        public double SubTotal => RequestedQuantity * Rate;
+        public decimal SubTotal => RequestedQuantity * Rate;
     }
 }
