+++ Models/DTO/ViewModels/SelectList/DeliveryService/DeliveryServiceVendor_SLM.cs
@@ -4,7 +4,7 @@
     public class DeliveryServiceVendor_SLM : SelectListModel
     {
         public bool IsSelf { get; set; }
-        public double ServiceCharges { get; set; }
+        public decimal ServiceCharges { get; set; }
         public bool ChargesInPercent { get; set; }
     }
 }
