+++ POS_API/Data/SalesOrderItemModifiers.cs
@@ -8,8 +8,8 @@ namespace POS_API.Data
         public int Id { get; set; }
         public int OrderItemId { get; set; }
         public int? ModifierId { get; set; }
-        public double? Quantity { get; set; }
-        public double? Charges { get; set; }
+        public decimal? Quantity { get; set; }
+        public decimal? Charges { get; set; }
         public int? BranchId { get; set; }
         public int CompanyId { get; set; }
         public int? Status { get; set; }
